fix(vulkan,rendering,editor,shader): Intel driver halt, vkQueueSubmit2, uint32 material maps (#773)
- Halt engine at startup when VK_DRIVER_ID_INTEL_PROPRIETARY_WINDOWS is detected:
  fences/semaphores signal before GPU execution completes on Intel HD/UHD Windows —
  no reliable Vulkan workaround exists. Clear user message directs to driver update
  or future D3D12 backend. Intel Arc (newer driver) and Linux ANV unaffected.
  Tracks as issue #775.

- Replace uint64_t material map fields with uint32_t in MaterialData (GLSL) and
  MeshMaterial (C++): removes GL_EXT_shader_explicit_arithmetic_types_int64 /
  shaderInt64 capability from GBuffer shader — Intel Iris Plus and other drivers
  that lack shaderInt64 now compile the pipeline successfully.

- Migrate all timeline semaphore submits from vkQueueSubmit+VkTimelineSemaphoreSubmitInfo
  to vkQueueSubmit2+VkSemaphoreSubmitInfo. Enable VkPhysicalDeviceSynchronization2Features.
  Fix type correctness: stage flag fields use VkPipelineStageFlags2. Remove
  ASSERT_TIMELINE_MONOTONIC and all 6 call sites.

- Deferred texture upload: texture async ops go into 1-frame SPSC deferred queues.
  Present() runs before SubmitAsyncUploads() so the same-frame GPU stall on drag-drop
  is eliminated. RequestDeferredDescriptorUpdate writes fallback then real image from
  the render thread to avoid concurrent vkUpdateDescriptorSets.

- Async drag-drop mesh deserialization: DeserializeMeshAssetFile and
  IngestMaterialFromUUID moved to ThreadPool worker with a dedicated ArenaAllocator
  (4x file size + 8 MB). Main-thread callback handles IngestMesh + actor create.

- Gate batch staging retirement on RenderTimeline (not m_batch_timeline): submit_1
  waits for batch before signalling RenderTimeline, guaranteeing GPU completion
  before staging buffers are freed. Submit batch immediately before Present() so
  geometry is always ready when render commands execute.

- Fix same-session texture bug: AppRenderPipeline::RenderScene used GetAsset<AssetMaterial>
  which returns null when VFSScanner marks the registry record Stale. Changed to
  UUIDToMaterialSlot.find() which is always authoritative after IngestMaterial.

+++ Resources/Shaders/g_buffer.frag
@@ -26,7 +26,7 @@ void main()
 
     if (material.AlbedoMap < INVALID_MAP_HANDLE)
     {
-        uint texId        = uint(material.AlbedoMap);
+        uint texId        = material.AlbedoMap;
         vec4 albedoSample = texture(sampler2D(TextureArray[nonuniformEXT(texId)], LinearWrapSampler), TexCoord);
         albedo            = albedoSample.rgb;
         alpha             = albedoSample.a;
@@ -37,7 +37,7 @@ void main()
 
     if (material.NormalMap < INVALID_MAP_HANDLE)
     {
-        uint texId        = uint(material.NormalMap);
+        uint texId        = material.NormalMap;
         vec3 normalSample = texture(sampler2D(TextureArray[nonuniformEXT(texId)], LinearWrapSampler), TexCoord).rgb;
         normal            = perturbNormal(normalize(WorldNormal), WorldPos, normalSample, TexCoord);
     }
@@ -45,7 +45,7 @@ void main()
     if (material.SpecularMap < INVALID_MAP_HANDLE)
     {
         // glTF metallicRoughnessTexture: R=occlusion, G=roughness, B=metallic
-        uint texId = uint(material.SpecularMap);
+        uint texId = material.SpecularMap;
         vec3 orm   = texture(sampler2D(TextureArray[nonuniformEXT(texId)], LinearWrapSampler), TexCoord).rgb;
         ao         = orm.r;
         roughness  = orm.g;
@@ -54,7 +54,7 @@ void main()
 
     if (material.EmissiveMap < INVALID_MAP_HANDLE)
     {
-        uint texId = uint(material.EmissiveMap);
+        uint texId = material.EmissiveMap;
         emissive   = texture(sampler2D(TextureArray[nonuniformEXT(texId)], LinearWrapSampler), TexCoord).r;
     }
